extract logic from main
