+++ lib/components/controls.js
@@ -21,6 +21,6 @@ module.exports = (activeFile) => {
 
 const saveFileButton = (activeFile) => {
   if (activeFile) {
-    return <button className="controls-open-file" onClick={() => saveFile(activeFile.fileName, activeFile.content, activeFile.id)}>Save Current File</button>;
+    return <button className="controls-open-file" onClick={() => saveFile(activeFile.fileName, activeFile.content, activeFile.id, currentWindow)}>Save Current File</button>;
   }
 };
