+++ lib/components/file-list-item.js
@@ -1,6 +1,7 @@
 import React from 'react';
 import store from '../store';
 import path from 'path';
+import truncate from '../truncate';
 
 const FileListItem = (file) => {
   const content = truncate(file.content, 25);
@@ -13,11 +14,6 @@ const FileListItem = (file) => {
       <div className="file-list-item-content">{content}</div>
     </div>
   );
-};
-
-const truncate = (text, maxLength) => {
-  if (text.length < maxLength) { return text; }
-  return text.slice(0, maxLength) + '...';
 };
 
 module.exports = FileListItem;
