+++ lib/main.js
@@ -15,57 +15,9 @@ app.on('ready', () => {
 });
 
 app.on('window-all-closed', function(event) {
-  console.log(event);
   app.quit();
 });
 
-const saveFile = (fileName, content, id) => {
-  if (fileName === 'New Note') fileName = `~/Documents/NewNote`;
-  dialog.showSaveDialog(
-      {defaultPath: fileName},
-    function (newFileName){
-      if (newFileName === undefined) return;
-      fs.writeFile(newFileName, content, 'utf8', (err) => {
-        if (err) throw err;
-        mainWindow.webContents.send('file-saved', newFileName, id);
-        dialog.showMessageBox({ message: "That note has most definitely been saved!",
-        buttons: ["Fantastic"]});
-      });
-  });
-};
-
-const saveFileOnClose = (fileName, content, id) => {
-
-  if (fileName === "New Note") {
-    let path = app.getPath('documents');
-    fileName =  `${path}/NewNote${id}`;
-    fs.writeFile(fileName, content, 'utf8', (err) => {
-      return { fileName, content, id };
-    });
-  }
-  fs.writeFile(fileName, content, 'utf8', (err) => {
-  });
-  return { fileName, content, id };
-};
-
-function closeProcedure(callback) {
-  let filesLastOpened = JSON.parse(localStorage.getItem('files')) || [];
-  let files = JSON.stringify(saveAllFiles(filesLastOpened));
-  localStorage.setItem('files', files);
-  if (typeof callback === 'function') {
-    callback();
-  }
-}
-
-const saveAllFiles = (files) => {
-  return files.map((file) => {
-     return Object.assign(file, saveFileOnClose(file.fileName, file.content, file.id));
-  });
-};
-
-
 exports.openFile = require('./open-file');
-exports.saveFile = saveFile;
-exports.saveFileOnClose = saveFileOnClose;
-exports.saveAllFiles = saveAllFiles;
-exports.closeProcedure = closeProcedure;
+exports.saveFile = require('./save-file');
+exports.saveAllFiles = require('./save-all-files');
