+++ lib/truncate.js
@@ -1,0 +1,6 @@
+
+
+module.exports =(text, maxLength) => {
+  if (text.length < maxLength) { return text; }
+  return text.slice(0, maxLength) + '...';
+};
